More explicit definition of available resources
The previous variant tried to ensure that everything was specified in an
abstract and disconnected manner.  Although the abstraction may well be
a good idea, the disconnection from the control-flow doesn't sound like
a good idea.  I prefer to know what is going on, which is the strategy
we're going for in this commit.