Support and implement update call
With this commit we have basic suppor for update calls.  The update is
always the full object, which may not be exactly what we want.  Also,
when something goes haywire, the original content is removed, rather
than us keeping the old content around.  We shouldfix these problems in
the near future.