Use special strings for escaping
The previous method was rather error-prone in the way it tried to export
content.  The translations had to be specified manually.  The new method
is not great, in that it still requires us to specify the types (and it
will not error when the type was not specified), but it's a lot better.