Implement GET to return complete instance
The GET request is a first draft of what it should become.  So far, it
contains the contents as strings.  There should be more extensive
support for rendering the results of a user's request.

This commit consists of an evolved version of the GET request.  The
initial implementation responded with the raw query results.