Parse input values through smarter parser
Input values are not always strings.  With this commit, we allow the
input values to consist of any json input, which will then be parsed to
contain the right type of content.  It would be possible to check the
input types when they are received, or by the description of the
resource.