Add system for interpreting user input
User input may be of various types.  Currently, we don't enforce any
specifics, however we should interpret the content the user provides in
the correct way.  Most important in this regard is the interpretation of
a URL.  This case in now handled by using the input types specified in
the resource description.