Allow users to specify links and host location
Users can now specify the links between various instances (not tested)
and can specify on which path the resource should be hosted.  It turns
out we need to generate quite a few request paths, hence it makes sense
to declaratively specify these paths.  This will later ensure we can
handle the links to other resource correctly.