Return 404 with correct content on wrong GET
A GET request which returns incorrect content should return a 404
response.  This is now handled by the system.  I suppose an issue may
still arize when a resource does not have all fields filled in.  For
now, we should handle these cases in migrations.  In the future, it
should be handled by sending the right query to the triple-store.