Add abstraction to construct not-found response
The 404 not found response is probably going to be a response which is
used more often.  An abstraction may help us tackle this case in a more
abstract manner.  This commit introduces the response-not-found method
which sends a 404 response with :null as primary data.  The response can
be extended if necessary.