Cope with incorrect Accept header in request
When the request does not contain the application/vndi.api+json request
header, we should return a 406 Not Acceptable with some content
describing the error.  We now throw this  error as required by the
spec.  For as much as we'd care, we'd allow this to be used as pure
application/json too, but the spec seems to force us (or at best
hint us not to) allow that.  Kind-of odd, but I can defend the point.