Simplify definition of simple conditions
We specify multiple conditions whose main content is the description and
the type.  Dispatching is handled by the type of the condition, what we
return to the user is handled by the description.  We shouldn't repeat
such code, it will fail un us when we make typos.  In this commit we've
constructed a general condition to be placed in the hierarchy above the
actual conditions.