Handle mismatch of request type
Users are obligated to supply the type of the object which should be
created, regardless of the path.  This means there may be a mismatch
between the path on which the request is made (and to what resource the
instance should belong due to it), ad the supplied type.  We now verify
the type is supplied correctly (as required by the jsonapi
specification) and return the correct 409 Conflict if it is not the
same.