Supply location header and self link on create
When we create a resource, we should supply the location header and a
self link.  Neither seems to be enforced, though it seems unlikely that
a client can correctly construct the path to the new resource in all
cases.  We therefore supply both.  As per spec, they must contain the
same content if both are supplied.