Abstract construction of resource item path
Where information on an item can be fetched can be derived from the
resource itself.  This requires stitching some content together so we
get the correct link.  The link might change over time, as we may start
understanding more information on where the frontend is hosting us.  For
this reason we are abstracting the creation of these links.  Later on,
this will give us a single place to alter the contents.