Construct PATCH request based on POST and PUT
The PATCH request will replace the current PUT request.  In this commit
we're only supporting the error handling that was supported by the POST
request.  We're also ignoring the fact that not the whole request may
have been supplied (the PATH object behaves much like a json
merge-patch (RFC 7386), or so it seems.