Cope with mismatching id in PATCH
The PATCH request must be made to a particular path.  The request itself
must contain both the type and the id.  If the id of supplied document
does not match the id for the specified path, we're not sure which we
should abide to.  This commit tackles that specific case.