Add supporting method to find a slot by json key
Our initial vision was that clients would supply the whole document on
each revision.  The jsonapi dictates differently with the PATCH
request only having to take the arguments which the client supplied into
account.  This may be a somewhat simpler and cleaner model in the long
run, but it requires us to access the descriptions of the keys for each
of the supplied keys, rather than for all keys.  The new getter method
allows us to easily fetch the content in a way that matches our intended
use.