Correct race condition of DELETE joined with LIST
When a DELETE and a LIST call are joined together and rapidly fired
after each other, it may well be that the LIST call will try to retrieve
an inexistant object.  We are now catching the case in which this
happens.  This is a real issue when using the json-api in a linked data
application.