Frameworks assume advised jsonapi type naming
The type naming should be up to the server.  However, the library we're
currently using on the Ember side seems to depend on the idea that the
type is the same as the first portion of the URL to fetch the resource
from.  As we're trying to be accepting to libraries, we'll join in with
this naming.

Note that I'm personally not very convenced that this is going to be a
good way of moving forward.  Then again, that might be exactly what
jsonapi is all about ;)