POC of partial PATCH on relation
This is a partial implementation of the link to a has-one instance.
It currently doesn't abide the use of inverse links.  However, the
specification does set the link as expected.

The fact that this is an inverse relation is simply skipped for now.  We
should disallow inverse relations to be set on the non-setting side, as
that may cause confusion.