Add support for deleting has-one links
In order to remove a has-one link, a PATCH request with null as the
content of the primary data object should be sent.  This has been
implemented by offering two paths in the single method responsible for
updating the has-one relationshisp.  It may be that this code will be
cut into pieces for supporting the has-many relationships if that would
proove to offer more code-reuse.