Add support for listing has-one relations
The GET request should show the content for the link containing the
objects which match the supplied request.  How that object should be
fetched is yet another thing, dependent on naming, however the type and
id is returned.  This is a has-one call, so the results are either
specified in the object itself, or by sending null as the primary data
of the result.