Add support for POST on relation
The POST call on a relation adds resources to the set of resources which
should be connected.  Support is fairly trivial as it's in the same
lines as the other methods which are offered.  We can probably benefit
from some code reuse between the PATCH call and the POST call here, but
that's deemed to be future work for now.