Allow listing of has-many resource
We already had support for showing the result of a has-one relation.
This commit adds support for the has-many relation which goes
hand-in-hand with the has-one relation.  I'm not very sure whether or
not this same support should be given on the relation object also.  I
suppose that will be clarified in one of the future versions of the
json-api specification.