Allow listing on relation resource and self link
It seems the ember addon we're using performs a GET request on the link
which contains the links section, rather than on the simplified one.
I'm not sure whether this is correct use or not, but it does not seem to
hurt us too much if we allow both to be queried for now.  Perhaps later
use will clarify an inconsistency on our end, or perhaps the future will
show us that this functionality is not necessary anymore.