Differentiate json-property and sparql-var + config
We incorrectly used the json-property-name as a basis for our SPARQL
variables.  This is not what the user should expect.  We have now
altered this so both are split.  In the same go, wa have made the syntax
for the attributes configurable.  You can now choose to have either
dasherized attributes, or camelcased attributes on an
application-level.  We need dasherized attributes for the plugin we are
currently using.