Renaming package to mu-cl-resources
The package renaming indicates the new usage of this library.  We now
indicate the name when being used inside the lisp world.

In further steps, which should be taken after a partial cleanup, the
packages.lisp file will be further adapted so it exports the correct
symbols.  That approach will allow us to import mu-cl-resources into a
new project and use it from there on.

For the case of mu.semte.ch, we will likely import a domain.lisp file
for the configuration of mu.semte.ch and work our way from there on.
This would not create a separate package for a specific project, but
works on Docker's premise that a single service does not need to care
about non-dependent other services.