Introducing ld-property-list for links and using
The ld-property-list returns the list as it should be used in a sparql
query.  This includes the indication of what constitutes an inverse
link.  Much of the code used custom logic to see whether an s-inv
wrapping should be added or not.  This issue has been resolved.  As an
added bonus, the cases which used to use s-inv already are now prepared
for relation paths, rather than a single relation.