Abstracting common pattern of find-resource-by-name
The pattern searches for the resource for a given link.  This is done by
getting the resource-name of the link and then finding the resource for
that specific name.  This should work correctly, unless the user has not
specified that resource.  That case is now also covered by sending off a
configuration-error to the user.