Manhandling deletion of resource
The deletion of a resource assumed all properties existed on the model.
As values may be optional, this is an incorrect assumption.  The
abstractions which have been put in place earlier on do not easily help
for this case.  Hence an inline solution has been constructed, ensuring
the right triples will be removed.  The previous approach with
sparql-delete-triples can (and should) be extended to support this new
form of removing triples.