Also support xsd:float
The list of types which are supported by mu-cl-resources is currently
constrained by what mu-cl-resources would put in the store itself.  As
mu-cl-resources is currently being used in a somewhat broader setting,
we will need to include more types.

There should be a more extensible way for adding types.  For now, this
will suffice.