Abstract the whole fetching of a paginated response
The paginated response returning is now fully abstracted.  Only the
graph pattern needs to be supplied.

In further commits we should abstract this further, ensuring the code
for finding such paginated responses is nice to read.  From there on, we
should be able to use this same code for paginating the relationships.