Add support for importing typed xsd:string
Typed strings are a bit dull, however that seems to be the way
UnifiedViews writes its content.  We will therefore support the
importing of these values for now.  Support for reading these strings
does raise the issue of being able to write typed strings through the
JSONAPI abstraction.  Something that is not supported now.