Removing the attributes collection for identity filter
The filter on id has a search path which ends with the resource itself.
Hence the search string provided by property-path-for-filter-components
will search for anything related, which has a mu:uuid.  That will
obviously not work in the vast majority of the cases.  In others, it'll
not mean what we intended it to mean.  This is spagetti code.  However,
as the whole section on building the filter query needs to be redone and
the deadlines are thight, we don't consider it to be the worst at this
stage.