Interpret null attribute on create as non-existant
The JSONAPI spec doesn't explicit whether or not null values are allowed
on a create, nor what their effect would be.  They were interpreted as
false values, which is most likely not what was expected.  We now
explicitly allow null values and we interpret them as the attribute
being non-existant.  These are the same as update semantics.