Enhance resource model with knowledge of its name
The resource can be found under a specific key as a name.  This key was
not known by the resource itself, which made for awkward wiring to be
done.  This wiring has now been altered so that the resource also knows
its name.  This feature is expected to be handy during development of
new features.