Basic POC showing included will be possible
This POC simply returns the id and type of the items which are to be
included.  However, this shows that we will be able to gather the
necessary content.  The building blocks are there, but we need a *much*
cleaner integration so we can keep it under control as it grows further.