Carve relieve-relation-items out of show-relation-call
This is only the implementation of the has-one-link, as the has-many
link can't supply the necessary contents.  This call will further be
altered so it only returns the ID and type of the resulting item.  But
in order to make that pretty, we have some other work to do first :)