Refactoring retrieve-item to only yield data content
Creation of the links portion is already partially abstracted.  It makes
more sense to yield the data and leave the wrapping to the consuming
entities.  This should make integration easier in the near future.

With this commit we've also removed the inclusion of the id on the
attributes section on creation.  This was once a requirement by jsonapi,
but that doesn't seem to be the case anymore.  Yay for consistency.