Allow inclusion of data object which matches item
If an item has related sources, then its contents should be included in
the related data item.  This first step can now occur with this
implementation.  In order to try it, toggle the last nil of
build-relationships-object to t.  All links will have a reference to the
data that's included.