Push objects to inlcude up to show-call
All the wiring has been done to push the objects which need to be
included in a request all the way up to the show-call.  This is a first
step towards getting the included objects inside our response.

At this stage, it seems that a two-tiered setup is going to be nicer
than pushing the items all the way up.  There are edge-cases in which an
item may be available through multiple relations.  A simple solution
would be to rerender the item when that happens.  However, this in
itself yields another issue.  There is a possibiltiy that an item is
available through more than one included path.  If that is the case,
then the expansions indicating what should be included /in/ said item,
may need to be extended.  You can re-render the item at this point, done
deal.  Well, not really... if the item is inside the primary response,
then it becomes a tad more complicated.  The current way of implementing
this whole shebang would still make sense, and it'll be the quickest
win, but I'm going to study whether building the data-model up front and
rendering that model afterwards won't yield us a higher quality
implementation.  This would be a good target on the long run anyway.