Code documentation with comments and tmp fixes
The inclusion of the related resources has been disabled by default.
The current implementation is in flux of additional features but should
be stable in terms of working functionality nonetheless.

We choose to push some bugfixes on top of this branch instead of on a
different branch to keep an understandable state.