Make retrieve-item and show-call cope with included
retrieve-item now accepts extra included items and will process them
through item-spec-to-jsown (which will need to cope with extended
information in the item-spec in the future.  show-call assumes this
exists and operates as necessary.

a related call, augment-data-with-attached-info has been stubbed.  this
call will have to calculate all inclusions and return them.  the
item-specs which are returned will also be augmented with information
regarding the relationships to include (and what the content of those
relationships is).