Rework allowing included for show-call
This allows the show-call to have an included portion.  This included
portion is shown to the user in a semi-correct fashion.  The
relationship objects don't contain a data attribute with the correct
type+uuids yet.  However, cycles should be allowed.

This implementation is also rather bad for listing calls, but those will
also start following the structure supplied here in the near future.