Tag version 1.4.0
New features:
- Implementation of included

Fixes since 1.3.0:
- Completed and updated exmaple of format in domain.lisp

Notes:

We have made the (unsupported) use of xsd:string typed strings as UUIDs
configurable.  See *allow-xsd-in-uuids* in configuration/settings.lisp.
This is mostly meant for measuring performance impact in realistic
situations.

Much code has been reworked in order to allow for a different internal
model when calculating requests.  In this new model, the items which
will need to be returned are calculated first (type, uuid, and
relationships to be included).  After that happened, the results are
sent to the underlying server.  This model should also make future
extensions simpler.