Allow sparse field sets to be requested
The implementation seems to be correct in the way it is now.  However,
it is a tad odd with the JSONAPI specification.  There's a statement
which says that, if there's a fields parameter for a type of resource,
that no other must be returned.  That's a bit of a bummer in case you've
requested related resources to be included.  I was inclined to include
the relationship in that case, but JSONAPI seems to decide differently.