Fix updating of partially known resources
If you intend to update a resource which is partially known (say, the
title is known, but the description isn't) and you sent a PACH so both
the title would be changed, and the description would be supplied, the
system did not correctly remove the currently known title from the
database.  Hence two values would exist at the end.