Add basic supporting code for checking access rights
The code presented here contains the a priliminary query which can be
used to check whether or not we have the right access rights.  The
structure of this query should be sound, but the specific predicates
will need to be discussed. Prefixes have been added to repository.lisp
to reflect the currently used predicates.

The supporting code alo contains some helpers which may be handy when
implementing the security layer in the call-implementation layer.

Some of this functionality may change as we get more experience with
this form of cross-cutting functionality.