Catch access-denied error in call-specification
The access-denied error is now caught in the call-specification.  This
ensures that the user is informed correctly when the server catches this
sort of exception.

We should streamline the catching of errors in the future.  There's way
too much code duplication going on there.  Most errors are caught in a
standardised way.