Tag version 1.7.0
New features:
- First draft of authorization
- Allow empty language-string-set to be specified
- Don't return multi-value properties when no values available
- Allow null for indication of empty relationship
- Add support for xsd:gYear by :g-year

Fixes since 1.6.0:
-
- Updating partially known sources
- Fix failures in relationships due to incomplete refactor

Notes:
The main feature here is a draft for resource authorization.  A bunch of
minor features and fixes sneaked into this release which pushes us to
release the authorization overly early.  This means that the
authorization as supplied today can be used for testing the logical
functionality, but that changes will happen in terms of which predicates
are going to be used to indicate the authorization.