First implementation of a callbacks system
The system allows us to define callabacks with any specification. Some
functional support is offered, sugercoated by macrology.  The approach
taken here is fairly easy to read.  We are aiming for code like the
following to work as expected:

(before (:create catalog)           ; this list contains the matching specs
    (input item-spec)               ; these are your received variables/arguments
  (declare (ignore input))          ; tell the compiler you don't care about input
  (error 'access-denied             ; don't allow the user to create, because fuck 'em!
         :operation :create
         :resource (resource item-spec)
         :id (uuid item-spec)
         :type (resource-name (resource item-spec))))

Before this implementation is complete, we will need to add the
necessary code to the implementation of the calls.

We looked around for callback systems which existed.  But in the light
of our beloved NIH syndrome, we found them not to be a perfect match for
our case.  The amount of syntax we'd need to lay over them was similar
to the amount of code that's written here.  This initial implementation
may evolve to suit our needs in the future.