Simplify sparql-pattern-filter-string
The method is becoming lengthier.  It will currently contain all cases
of filtering.  As this method becomes more and more complex, there are
cleaning steps.  Currently, we're cleaning this method to be clearer to
read.  If this becomes too complex, we'll implement some supporting code
so we can base ourselves on pattern-like matching.