Try out alternative filter[:concept:var] syntax
The new syntax starts with a colon (:) followed by a type of filter,
followed by the arguments of that filter.  In the case of [id], this has
become [:id:] (both are supported for now).  In the case of an exact
match for the tag attribute, we use [:exact:tag].  The syntax has not
been finalized yet, but this is a good candidate.