Deprecate use of filter[id] in favour of filter[:id:]
Altough the new syntax is not finalized yet, we should transition to a
special form for all special functionality.  id is not your regular key,
perhaps it shouldn't be treated as that.

There is something to say for allowing id to be non-special, as it is
also a reserved attribute name in the jsonapi specification.