Add feature system with example include-uri feature
The feature system allows to specify features for a resource.  The
features which a resource has can be used to customise the logic
regarding the resources.  This way of working could allow us to
customise the response, for instance.  The current example is exposing
the uri of the resource in the attributes of the response.

Note: this whole system is considered experimental.  It's a path which
we're testing and which may be removed in the near future.