Support for :string-set type
A string-set represents a set of regular strings as a value, rather than
a single string.  This allows us to refer to a bunch of strings as a
value, rather than a single string.

When looking at the implementation, it is clear that the current way of
representing property types is suboptimal.  In terms of the combination,
one could imagine that a set of any primitive could hold.  The current
implementation doesn't scale in this regard.