Tag version 1.10.4
Fixes:
- Fix broken cl-fuseki

Notes:
During development on the whole mu-cl-resources ecosystem in order to
adapt it to ABCL running with a different webserver, a bunch of
debugging statements were added.  These statements can cause errors.
This debugging statement is a result of that.  It might be that we
encounter others, but I couldn't immediately see them in the current
sources.