Adding feature to disable default pagination
When the user doesn't request pagination, PoCs may prefer not to have
pagination in the API.  The no-pagination-defaults feature can be used
in this case.  It is not an advised feature, and it should raise alert
whenever you see it in a domain.lisp file, but it may alleviate some
growing pains.