Reset cache for link on resource update
The updated resource may appear in a relationship.  We err on the side
of fewer keys and more cache clears, by clearing the relationships
whenever a resource of that type has been updated.  Optimisations in
which we list the specific resources are plausible extensions.

This is an experimental fix for an experimental feature.