Less chance for duplicate cache/clear keys
In the previous implementation it was feasible that a certain key was
sent to the cache multiple times.  This is still possible, but the
chances are vastly lower.  We now use the json serialization as the
hash-key fort he table containing the cache/clear keys.  Hence, if the
order of the attributes of the json objects is fixed, this will work
without duplicates.