Eager clearing of inverse relationships
Due to the currently limited knowledge for the clearing of
relationships, we are eagerly clearing complete inverse relationships.
This approach makes the cache feasible for real-world scenarios with
limited downsides.

We expect improvements to the cache to be made over time.  Our goal
would then be to query the store for the current state and explicitly
clear our inverse relationships.