Correct self for relationship calls
When relationships were shown, the self was the core of the
relationship, not including any sorting etc.  This seems to be
incorrect as per spec.  We now yield all query parameters.

This current approach may not be completely right either.  The spec
says that it should be "the link that generated the current response
document.".  To my understanding this may (but is not required to)
contain query parameters with default values, or query parameters
which otherwise have no particular impact on the response (eg:
unsupported parameters).  I'm currently not sure what would be
preferred in these situations.