Do not include duplicate relationships in included
When fetching items with included relationships (using ?include=), you
could run into the issue that an item of the relationship was included
multiple times.  This can happen when the triplestore can find
multiple matching patterns for a single returned ?uuid.  This has been
solved by adding a DISTINCT marker.

The DISTINCT marker has been added in the latest commit of
mu-cl-support.  We are therefore updating mu-cl-support.