Correct SPARQL query when sorting and filtering
SELECT queries which have a GROUP-BY should list all the selected
variables in the GROUP-BY as per SPARQL spec, unless they are grouped
in a way.  For variables on which we intend to sort, we should have a
structured way of grouping them.  Currently, we select the MAX value,
which uses the same sorting as ORDER-BY.  This is not a superb way to
sort, but we expect multiple results to most often contain a single
value.  In that case it doesn't matter which one select, as long as
the query is valid.  Results are expected to be more stable than they
were before this change occurred.