Refactor property-path-for-filter-components to expose more info
This method is used for special filters.  It used to expose the SPARQL
property path for the series of keys in the filter coming from the
URL.  By splitting this method in two, it became more legible.
Supplying extra information to the caller makes it easier to provide
rich search interfaces.