Ensure :exact: is first word in search key
With only :exact: as a special key, it doesn't matter whether the keys
was matched as the first word, or as a later word.  The net effect
would -at worst- be an error.  However, with multiple of these coming
into place, :exact: may be repeated elsewhere under a different
meaning.  We now check this to be the first word.