Output numbers as xsd:decimal in queries
This should be equivalent to writing the number without quotes in the
SPARQL query.  However, this approach allows us to reuse the same
logic when building up the SPARQL constraint.  In the case of the
constraint, the number arrives as a string parameter, rather than a
real number.  There are two approaches possible.  One approach
dictates we convert the string to a number based on its type.  The
other approach is to allow the :number type to be supplied by either a
number or a string.  We opted for the second option as it allows us to
maintain one code-base.  Plus, it allows us to be more forgiving to
input.