Correct writing of ratio's to the triplestore
JSOWN interprets float's as precise as it can by using a decimal.  0.2
would be therefore be injected into a query as "1/5"^^xsd:decimal.
Writing out this number will generate something the triplestore
doesn't understand.  As such we now write numbers to the query either
as integers or as floating point numbers.