Use a typed literal rather than a filter clause
A filter clause will force a table scan when executing the query.
Replacing this with a typed literal match (when the option is set in settings.lisp)
allows the database to use an indexed lookup. The result is a speedup of
several orders of magnitude.

mu-cl-resource was also not taking the allow-xsd-in-uuids flag into account
when parsing jsonapi filter clauses