Streamline cache settings
The cache settings were not stored in the settings.lisp and contained
over-agressive defaults.

Two source files had the defparameter which allowed for configuring
the caching inline with the sources, rather than in the
settings.lisp file.  These configurations have been moved.

The cache-count-queries setting defaulted to t, which would yield
inconsistencies when multiple services would write to the database.
It now defaults to nil.