Add *max-group-sorted-properties* option to drop MAX in SELECT
Properties may have multiple values in the triplestore.  When
searching for UUIDs, we therefore group the results per uuid to ensure
we don't have duplicate values.  Sorting may also occur when fetching
these lists.  In order to ensure the sorting is consistent, we take
the maximum value from the set of values which the property may have.
This seemingly yields problems in Virtuoso.  Hence there's now the
option to disable this functionality and assume there are no duplicate
values or assume the triplestore will do a sane job anyways.  We
intend to drop this feature if/when triplestores get more mature about
this.