Scope count cache by access rights
The access rights of the user are now taken into account when
calculating the count cache.  There are various optimizations possible
here, as the clearing of the cache keys is overly aggressive.  Other
cases will be upgraded transparently as we learn to cope with the
mu-auth-allowed-groups and mu-auth-used-groups, but this case could
benefit from extra opmitizations.

The current implementation is on the safe side, hence we shouldn't
expect issues in the near future.