Clean up created threads
Calculation of included items is now ran in parallel through
lparallel.  We create a fresh set of 8 threads when this case occurs.
However, there was no cleanup of these threads.  This results in many
threads being created and not being cleaned up.  This obviously makes
the system unhappy.

The new implementation ensures the threads are cleaned up once the
included items are calculated.  We could choose to optimize the server
load by creating a shared pool fo resources.  For now, we assume the
calls can cleanly run in parallel.