Add experimental cache-model-properties resource feature
The cache-model-properties resource feature allows caching of model
properties on a per-resource basis.  This means that you can set some
resources not to have their model properties cached, whilst others
would have such a cache.

Use :features '(cache-model-properties) in the define-resource.

This feature is EXPERIMENTAL.  We intend to merge this feature with
other features and/or to merge it over time.