PoC for cache clearing through delta messages
This PoC shows that we can clear the cache through changes received
from the Delta service.  The PoC has quite a few very rough edges
which must be tackled before putting this in the wild.  Nevertheless,
this is a working state, we should be able to refer back to it.