Use existing primitives for clearing caches
By using the common primitives we have much less chance for bugs.  In
fact, we seemingly squashed some unconfirmed bugs in the process.  The
applied approach has not been tested in any but the most minimal
cases.  The construction for coming to the right sources is rather
complex and should be simplified.