Ignore delta messages which don't cover the mu-cl-resources domain
The domain of the stack may be wider than the domain mu-cl-resources
knows of.  In order to cope with various stacks, mu-cl-resources now
accepts delta messages which it doesn't know about.  It will not clear
anything except for the class resource in that case.

Note that we're logging information on CASE E11 which we currently
assume may not get triggered.