Cache fetching of uri classes in single delta call
For a single delta call we now cache the classes of URIs which we have
discovered.  This should yield to some performance improvement,
depending on the type of load.

This functionality could be further optimized by working with a user
aware cache assuming that would start coping with sudo too.