Better interpretation for json value types
The json value types are now interpreted in a better fashion, meaning
primitive values that exist in json can now be entered as you'd expect
them to be entered in the json body.  There's subtle difference in how
you enter these in the query param, but feature requests seem to
indicate that people largely ignore the fact that they're entering a
string or a boolean, hence they expect the value to be interpreted
based on the slot type automagically.  We now do just that in more
cases.