Use top-level paths instead of resources
In discussions, it is slowly becoming nicer to use path as a top-level
element.  The `name` would be an intrinsic element in this case.
Overriding the name would still be an option if desired.

We've chosen to make both path as well as name optional keys in the
description of a resource.  Both of these default to the supplied key
in the top-level `paths` property.  Some discussion around the key
`paths` could now pop up as a more-or-less final point of decision.