First stab at inheritance
A first stab at inheritance.  Although this is most likely not
complete, this implementation does allow us to play around with the
general idea.

Adding inheritance may have repercussions in many small areas.  As
such, the best way to land this, will be to go over the various
aspects of mu-cl-resources and see how to best integrate these
features by avaluating the specific cases.