step towards moving to reified properties for resource types
the values we are going to be interested in will almost always be the
values from the combined hierarchical tree.  we should return those
where plausible.

this is a step in this direction.  we should have ld-properties of a
resource and the resource of an item-spec by now, but there are a few
more properties which should behave similarly.