Clear caches after updating the database to mitigate cache poisoning
There's a case in which updates are sent and read at the same time.
If timed correctly, this could poison the cache, making
mu-cl-resources believe the state had been updated, even when it was
not.  This commit clears the caches only after the state has been
written to the database.