mu-auth-allowed-groups improvements for locally calculated results
Some results can be fully calculated by mu-cl-resources.  In case the
request could be fully calculated by mu-cl-resources, no
mu-auth-allowed-groups were returned.  Shifting off this responsibility
to the database only works when we actually need the database.
mu-cl-resources can do without the database in some cases, leading to
results that could not be cached.

The solution is imperfect.  As we intend for the mu-auth-allowed-groups
to be extended with more well-understood mu-auth-used-groups, this
probably isn't the very worst.