Cope with non-existent UUIDs
When UUIDs don't exist, the error messages would try to fetch too much
information, leading to continued fetching of resources that would lead
to no results.  This has been cleaned out.  As we went along, it turns
out we were catching a particular error which is also caught higher-up.
We're now relying on the top-level catching as that has sufficient
control over the errors.