Add at least one required triple when fetching properties
The properties fetching only fetches the necessary properties.  This set
of properties now includes at least one required triple when adding
properties.

We've noticed Virtuoso barfs on us if we only request optional
properties.  A tad surprised we only notice this now so perhaps it
doesn't barf in all cases.  If it does, this should offer a solution in
it requesting at least one known triple (the uuid).

We could possibly make this more efficient by fetching the type, yet
with the inheritance system we're not sure only one type exists and we
may thus make the result set explode.  Hence the choice of using the
uuid.  There are downsides to using the mu:uuid also.  We currently
assume the mu prefix is what we expect it to be whilst it could
theorethically be overwritten with another value.  I suspect overriding
this would lead to many other errors and it would make more sense to
guard for it at bootup, rather than to make different assumptions.