Enhanced escaping of filtered values in query parameters
When a value is provided in a query parameter, it was not parsed
correctly.  In the past, values could be quoted, but that didn't lead to
the desired results either.  We are currently very lenient in what we
accept as an input and we parse the values as if they were json
properties if they are a primitive value.